backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
supabase
python-dotenv
celery
redis
python-multipart
python-magic-bin
email-validator
duckdb